Create Hacker-Rank flippingMatrix
C++

int flippingMatrix(vector<vector<int>> matrix) {
    //int rows = matrix.size();
    //int columns = matrix[0].size();
    int n = matrix.size()/2; //as this matrix is 2n x 2n.
    int largestSum = 0;
    
    // A B B A          A B C C B A
    // C D D C    [OR]  D F H H F D                 [So On...]
    // C D D C          E G I I G E
    // A B B A          E G I I G E
    //  2nx2n           D F H H F D
    // 2(2)x2(2)        A B C C B A
    //   4x4            2nx2n => 2(3)x2(3) => 6x6
    
    for(int i = 0; i < n; i++){
        for(int j = 0; j < n; j++){
            largestSum += max(matrix[i][j], max(matrix[i][matrix.size()-j-1], max(matrix[matrix.size()-i-1][j], matrix[matrix.size()-i-1][matrix.size()-j-1])));
        }
    }
    return largestSum;
}
This problem is like rubix cube but in a 2D format.

The key for this problem is to identify the poisitions of the potential mirroring numbers in this matrix that can be flipped with each of the numbers in the top-left nxn matrix.

As said by many, the problem statement is a bit misleading as we first get the thought of using reverse function from STL.

I tried a manual way of reversing in 6x6 matrix using pen & paper. Flipping/Reversing isn't necessarily needed. Rather, finding the largest number among the numbers in potential mirroring positions to the top-left nxn matrix numbers is enough. Finally Sum up all the largest numbers found.